Unify StatusHttpHandler and add git_info_sha information
