+++ src/meta/test/CMakeLists.txt
@@ -190,46 +190,6 @@ nebula_add_test(
 
 nebula_add_test(
     NAME
-        meta_http_status_test
-    SOURCES
-        MetaHttpStatusHandlerTest.cpp
-    OBJECTS
-        $<TARGET_OBJECTS:meta_http_handler>
-        $<TARGET_OBJECTS:meta_client>
-        $<TARGET_OBJECTS:stats_obj>
-        $<TARGET_OBJECTS:meta_service_handler>
-        $<TARGET_OBJECTS:storage_thrift_obj>
-        $<TARGET_OBJECTS:kvstore_obj>
-        $<TARGET_OBJECTS:meta_thrift_obj>
-        $<TARGET_OBJECTS:common_thrift_obj>
-        $<TARGET_OBJECTS:raftex_obj>
-        $<TARGET_OBJECTS:raftex_thrift_obj>
-        $<TARGET_OBJECTS:wal_obj>
-        $<TARGET_OBJECTS:thrift_obj>
-        $<TARGET_OBJECTS:ws_obj>
-        $<TARGET_OBJECTS:ws_common_obj>
-        $<TARGET_OBJECTS:hdfs_helper_obj>
-        $<TARGET_OBJECTS:http_client_obj>
-        $<TARGET_OBJECTS:base_obj>
-        $<TARGET_OBJECTS:fs_obj>
-        $<TARGET_OBJECTS:time_obj>
-        $<TARGET_OBJECTS:stats_obj>
-        $<TARGET_OBJECTS:network_obj>
-        $<TARGET_OBJECTS:thread_obj>
-        $<TARGET_OBJECTS:process_obj>
-        $<TARGET_OBJECTS:schema_obj>
-        $<TARGET_OBJECTS:gflags_man_obj>
-    LIBRARIES
-        proxygenhttpserver
-        proxygenlib
-        ${ROCKSDB_LIBRARIES}
-        ${THRIFT_LIBRARIES}
-        wangle
-        gtest
-)
-
-nebula_add_test(
-    NAME
         meta_http_download_test
     SOURCES
         MetaHttpDownloadHandlerTest.cpp
