+++ src/storage/test/CMakeLists.txt
@@ -133,29 +133,6 @@ nebula_add_test(
 
 nebula_add_test(
     NAME
-        storage_http_status_test
-    SOURCES
-        StorageHttpStatusHandlerTest.cpp
-    OBJECTS
-        $<TARGET_OBJECTS:storage_http_handler>
-        $<TARGET_OBJECTS:http_client_obj>
-        $<TARGET_OBJECTS:ws_obj>
-        $<TARGET_OBJECTS:stats_obj>
-        $<TARGET_OBJECTS:process_obj>
-        $<TARGET_OBJECTS:adHocSchema_obj>
-        $<TARGET_OBJECTS:meta_service_handler>
-        ${storage_test_deps}
-    LIBRARIES
-        proxygenhttpserver
-        proxygenlib
-        ${ROCKSDB_LIBRARIES}
-        ${THRIFT_LIBRARIES}
-        wangle
-        gtest
-)
-
-nebula_add_test(
-    NAME
         storage_http_admin_test
     SOURCES
         StorageHttpAdminHandlerTest.cpp
