+++ src/webservice/StatusHandler.cpp
@@ -1,0 +1,75 @@
+/* Copyright (c) 2019 vesoft inc. All rights reserved.
+ *
+ * This source code is licensed under Apache 2.0 License,
+ * attached with Common Clause Condition 1.0, found in the LICENSES directory.
+ */
+
+#include "webservice/StatusHandler.h"
+#include <proxygen/httpserver/RequestHandler.h>
+#include <proxygen/lib/http/ProxygenErrorEnum.h>
+#include <proxygen/httpserver/ResponseBuilder.h>
+
+namespace nebula {
+
+using proxygen::HTTPMessage;
+using proxygen::HTTPMethod;
+using proxygen::ProxygenError;
+using proxygen::UpgradeProtocol;
+using proxygen::ResponseBuilder;
+
+void StatusHandler::onRequest(std::unique_ptr<HTTPMessage> headers) noexcept {
+    if (headers->getMethod().value() != HTTPMethod::GET) {
+        // Unsupported method
+        err_ = HttpCode::E_UNSUPPORTED_METHOD;
+        return;
+    }
+}
+
+void StatusHandler::onBody(std::unique_ptr<folly::IOBuf>) noexcept {
+    // Do nothing, we only support GET
+}
+
+void StatusHandler::onEOM() noexcept {
+    switch (err_) {
+        case HttpCode::E_UNSUPPORTED_METHOD:
+            ResponseBuilder(downstream_)
+                .status(WebServiceUtils::to(HttpStatusCode::METHOD_NOT_ALLOWED),
+                        WebServiceUtils::toString(HttpStatusCode::METHOD_NOT_ALLOWED))
+                .sendWithEOM();
+            return;
+        default:
+            break;
+    }
+
+    folly::dynamic vals = getStatus();
+    ResponseBuilder(downstream_)
+        .status(WebServiceUtils::to(HttpStatusCode::OK),
+                WebServiceUtils::toString(HttpStatusCode::OK))
+        .body(folly::toJson(vals))
+        .sendWithEOM();
+}
+
+
+void StatusHandler::onUpgrade(UpgradeProtocol) noexcept {
+    // Do nothing
+}
+
+
+void StatusHandler::requestComplete() noexcept {
+    delete this;
+}
+
+
+void StatusHandler::onError(ProxygenError error) noexcept {
+    LOG(ERROR) << "Web service StorageHttpHandler got error: "
+               << proxygen::getErrorString(error);
+}
+
+folly::dynamic StatusHandler::getStatus() {
+    folly::dynamic json = folly::dynamic::object();
+    json["status"] = "running";
+    json["git_info_sha"] = NEBULA_STRINGIFY(GIT_INFO_SHA);
+    return json;
+}
+
+}  // namespace nebula
