+++ src/webservice/WebService.cpp
@@ -11,6 +11,7 @@
 #include "webservice/GetFlagsHandler.h"
 #include "webservice/SetFlagsHandler.h"
 #include "webservice/GetStatsHandler.h"
+#include "webservice/StatusHandler.h"
 
 DEFINE_int32(ws_http_port, 11000, "Port to listen on with HTTP protocol");
 DEFINE_int32(ws_h2_port, 11002, "Port to listen on with HTTP/2 protocol");
@@ -89,6 +90,12 @@ Status WebService::start() {
     });
     registerHandler("/get_stats", [] {
         return new GetStatsHandler();
+    });
+    registerHandler("/", [] {
+        return new StatusHandler();
+    });
+    registerHandler("/status", [] {
+        return new StatusHandler();
     });
 
     std::vector<HTTPServer::IPConfig> ips = {
